[IOTMID-80] Changes option values for device nodes
This commit changes the value that is stored on the combo box
for the known devices, storing the device id instead of its
position in the retrieved array.

This also changes the package's readme file with instructions on
how to install the package without having to resort to global
linkage, and the correct location so that the json that gets
exported from the ui contains the newly installed nodes.

Change-Id: I9d77fe21042786b26032e905e24b56e0163fe63a