[IOTMID-416][IOTMID-418] Adds email, geofence nodes
This moves the custom nodes implementation around, and adds
customized versions of two new nodes: email and geofence.

To have a properly configured image, it is important to run
npm install before creating the docker image. That way the
dependencies for the new nodes are properly installed under
node_modules, and node-red stops complaining about missing
deps when starting the process.

Change-Id: Id872553103a5e828e44692975347f24b72e3db43